requirements.txt
fastapi>=0.116,<1
uvicorn[standard]>=0.35,<1
sqlalchemy>=2.0,<3
psycopg[binary]>=3.2,<4
pydantic>=2.11,<3
python-multipart>=0.0.20,<1
redis>=6,<7
rq>=2.6,<3
google-api-python-client>=2.178,<3
google-auth>=2.40,<3
google-auth-httplib2>=0.2,<1
httpx>=0.28,<1
tenacity>=9,<10
pytest>=8.4,<9
